+++ CHANGELOG.md
@@ -32,6 +32,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Chore
 
+- Bump kotlin-reflect from 1.3.41 to 1.3.50 in /analysis
+- Bump json from 20180813 to 20190722 in /analysis
+- Bump rxjava from 2.2.9 to 2.2.12 in /analysis
+- Bump assertj-core from 3.12.2 to 3.13.2 in /analysis
+- Bump sonar-java-plugin from 5.12.1.17771 to 5.14.0.18788 in /analysis
+
 ## [1.33.0] - 2019-09-10
 
 ### Added